pyproject.toml
[project]
name = "coach-notifier"
version = "0.1.0"
description = ""
authors = [
    {name = "Chase Switzer",email = "chasehillig@yahoo.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "python-dotenv (>=1.1.1,<2.0.0)",
    "requests (>=2.32.4,<3.0.0)",
    "fastapi[standard] (>=0.116.0,<0.117.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
